Decoupled Training Botbase Removed Pickepet Timer from UI and created new one in TrainingManager.cs
