+++ Botbases/RSBot.Training/Container.cs
@@ -5,8 +5,6 @@ namespace RSBot.Training;
 
 internal static class Container
 {
-    private static Main _mainView;
-
     /// <summary>
     ///     Gets or sets the bot.
     /// </summary>
@@ -14,23 +12,6 @@ internal static class Container
     ///     The bot.
     /// </value>
     public static Botbase Bot { get; set; }
-
-    /// <summary>
-    ///     Gets or sets the view.
-    /// </summary>
-    /// <value>
-    ///     The view.
-    /// </value>
-    public static Main View
-    {
-        get
-        {
-            if (_mainView == null || _mainView.Disposing || _mainView.IsDisposed)
-                _mainView = new Main();
-
-            return _mainView;
-        }
-    }
 
     /// <summary>
     ///     Gets or sets the lock.
