+++ Botbases/RSBot.Training/Views/View.cs
@@ -1,0 +1,13 @@
+﻿namespace RSBot.Training.Views
+{
+    internal class View
+    {
+        /// <summary>
+        ///     Gets or sets the instance.
+        /// </summary>
+        /// <value>
+        ///     The instance.
+        /// </value>
+        public static Main Instance { get; } = new();
+    }
+}
